week5/requirements.txt
torch>=1.10
torchvision>=0.11
numpy
matplotlib
tqdm
Pillow
ultralytics